requirements.txt
requests>=2.31.0
pyyaml>=6.0.1
jinja2>=3.1.3
urllib3>=2.1.0
